Move IM/GM initialization to a C file, take care of signals
Previously, we were just telling IM and GM not not mess with Go signals.
However, since Go 1.6 the restrictions are tigher and we can no longer
rely on IM and GM to properly set up their signal handlers exactly how
we need them to be, so we now fix the handlers after calling IM/GM
initialization code.

Note that this creates a very small window when a signal arrival will
panic, since the handlers are incorrectly set up between the IM/GM
initialization and magick_fix_signals(), but it's the best we can do.

Fixes #37.